agents/strands/customer-support-assistant/dev-requirements.txt
# Development requirements for Customer Support Assistant
# Generated from pyproject.toml dependencies

# Core AgentCore dependencies
bedrock-agentcore>=0.1.0
bedrock-agentcore-starter-toolkit>=0.1.0

# AWS SDK
boto3>=1.39.7
botocore>=1.39.7

# CLI and utilities
click>=8.2.1

# Google APIs
google-api-python-client>=2.176.0
google-auth>=2.40.3

# Search and database
opensearch-py>=3.0.0

# Data processing
pandas>=2.3.1

# Configuration and serialization
pyyaml>=6.0.2

# AWS authentication
requests-aws4auth>=1.3.1

# Utilities
retrying>=1.4.0

# AI agents and tools
strands-agents==1.7.0
strands-agents-tools>=0.2.0

# Web interface
streamlit>=1.47.0
streamlit-cookies-controller>=0.0.4

# Additional development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0

#ollama
ollama